Add embedded infinity alter statement.
